+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+### Added
+- Add ExecutableIcon property on IGame interface / implementations
 
 
 ## [1.1.1] - 2022-27-08